raspberry/requirements.txt
fastapi
uvicorn
RPi.GPIO
pigpio
